move keyhandler
