+++ oidcdiscovery/oidcdiscovery_test.go
@@ -369,71 +369,6 @@ func TestNew(t *testing.T) {
 	}
 }
 
-func TestNewKeyHandler(t *testing.T) {
-	op := server.NewTesting(t)
-	issuer := op.GetURL(t)
-	discoveryUri := getDiscoveryUriFromIssuer(issuer)
-	jwksUri, err := getJwksUriFromDiscoveryUri(discoveryUri, 10*time.Millisecond)
-	require.NoError(t, err)
-
-	keyHandler, err := newKeyHandler(jwksUri, 10*time.Millisecond)
-	require.NoError(t, err)
-
-	keySet1 := keyHandler.getKeySet()
-	require.Equal(t, 1, keySet1.Len())
-
-	expectedKey1, ok := keySet1.Get(0)
-	require.True(t, ok)
-
-	token1 := op.GetToken(t)
-	keyID1, err := getKeyIDFromTokenString(token1.AccessToken)
-	require.NoError(t, err)
-
-	// Test valid key id
-	key1, err := keyHandler.getByKeyID(keyID1, false)
-	require.NoError(t, err)
-	require.Equal(t, expectedKey1, key1)
-
-	// Test invalid key id
-	_, err = keyHandler.getByKeyID("foo", false)
-	require.Error(t, err)
-
-	// Test with rotated keys
-	op.RotateKeys(t)
-
-	token2 := op.GetToken(t)
-	keyID2, err := getKeyIDFromTokenString(token2.AccessToken)
-	require.NoError(t, err)
-
-	key2, err := keyHandler.getByKeyID(keyID2, false)
-	require.NoError(t, err)
-
-	keySet2 := keyHandler.getKeySet()
-	require.Equal(t, 1, keySet2.Len())
-
-	expectedKey2, ok := keySet2.Get(0)
-	require.True(t, ok)
-
-	require.Equal(t, expectedKey2, key2)
-
-	// Test that old key doesn't match new key
-	require.NotEqual(t, key1, key2)
-
-	// Validate that error is returned when using fake jwks uri
-	_, err = newKeyHandler("http://foo.bar/baz", 10*time.Millisecond)
-	require.Error(t, err)
-
-	// Validate that error is returned when keys are rotated,
-	// new token with new key and jwks uri isn't accessible
-	op.RotateKeys(t)
-	token3 := op.GetToken(t)
-	keyID3, err := getKeyIDFromTokenString(token3.AccessToken)
-	require.NoError(t, err)
-	op.Close(t)
-	_, err = keyHandler.getByKeyID(keyID3, false)
-	require.Error(t, err)
-}
-
 func TestGetHeadersFromTokenString(t *testing.T) {
 	key, _ := testNewKey(t)
